requirements.txt
ultralytics>=8.0.0
opencv-python>=4.5.0
numpy>=1.24.0
pyyaml>=6.0
matplotlib>=3.7.0
torch>=2.0.0
